add f64/f32 features to kesko_physics (#127)
in order to switch between f32 and f64 for rapier the physic crate now reflect the same feature. the cut is not super clear. some conversion are needed in other creates and maybe more creates should have this feature in the future.

by default f64 is enabled.

closes: #125

+++ crates/kesko_core/src/cursor_tracking.rs
@@ -97,9 +97,9 @@ pub(crate) fn update_tracking_controller_system<T: Component + Default>(
                 let mut control_output = track.controller.act(pos_diff);
 
                 if let Some(mass) = multibody_mass {
-                    control_output *= mass.val;
+                    control_output *= mass.val as f32;
                 } else {
-                    control_output *= mass.val;
+                    control_output *= mass.val as f32;
                 }
 
                 force.vec = control_output;
